+++ src/local_scheduler/local_scheduler_algorithm.cc
@@ -405,21 +405,16 @@ void finish_killed_task(LocalSchedulerState *state,
   }
   /* Mark the task as done. */
   if (state->db != NULL) {
+    Task *task = Task_alloc(execution_spec, TASK_STATUS_DONE, get_db_client_id(state->db));
     #if !RAY_USE_NEW_GCS
-      Task *task = Task_alloc(execution_spec, TASK_STATUS_DONE,
-                              get_db_client_id(state->db));
       // In most cases, task_table_update would be appropriate, however, it is
       // possible in some cases that the task has not yet been added to the task
       // table (e.g., if it is an actor task that is queued locally because the
       // actor has not been created yet).
       task_table_add_task(state->db, task, NULL, NULL, NULL);
     #else
-      TaskSpec* spec = execution_spec.Spec();
-      auto data = MakeTaskTableData(execution_spec, get_db_client_id(state->db), SchedulingState_DONE);
-      RAY_CHECK_OK(state->gcs_client.task_table().Add(ray::JobID::nil(), TaskSpec_task_id(spec), data,
-                                                      [](gcs::AsyncGcsClient *client,
-                                                         const TaskID &id,
-                                                         std::shared_ptr<TaskTableDataT> data) {}));
+      RAY_CHECK_OK(TaskTableAdd(&state->gcs_client, task));
+      Task_free(task);
     #endif
   }
 }
@@ -535,16 +530,9 @@ void queue_actor_task(LocalSchedulerState *state,
       #if !RAY_USE_NEW_GCS
         task_table_update(state->db, task, NULL, NULL, NULL);
       #else
-        TaskExecutionSpec& execution_spec = *Task_task_execution_spec(task);
-        TaskSpec* spec = execution_spec.Spec();
-        auto data = MakeTaskTableData(execution_spec, Task_local_scheduler(task), SchedulingState_QUEUED);
-        RAY_CHECK_OK(state->gcs_client.task_table().Add(ray::JobID::nil(), TaskSpec_task_id(spec), data,
-                                                         [](gcs::AsyncGcsClient *client,
-                                                           const TaskID &id,
-                                                          std::shared_ptr<TaskTableDataT> data) {}));
+        RAY_CHECK_OK(TaskTableAdd(&state->gcs_client, task));
         Task_free(task);
       #endif
-      // TODO(pcm): update actor task
     } else {
       /* Otherwise, this is the first time the task has been seen in the
        * system (unless it's a resubmission of a previous task), so add the
@@ -552,11 +540,8 @@ void queue_actor_task(LocalSchedulerState *state,
       #if !RAY_USE_NEW_GCS
         task_table_add_task(state->db, task, NULL, NULL, NULL);
       #else
-        auto data = MakeTaskTableData(execution_spec, Task_local_scheduler(task), SchedulingState_QUEUED);
-        RAY_CHECK_OK(state->gcs_client.task_table().Add(ray::JobID::nil(), TaskSpec_task_id(spec), data,
-                                                        [](gcs::AsyncGcsClient *client,
-                                                           const TaskID &id,
-                                                          std::shared_ptr<TaskTableDataT> data) {}));
+        RAY_CHECK_OK(TaskTableAdd(&state->gcs_client, task));
+        Task_free(task);
       #endif
     }
   }
@@ -910,10 +895,9 @@ std::list<TaskExecutionSpec>::iterator queue_task(
   /* The task has been added to a local scheduler queue. Write the entry in the
    * task table to notify others that we have queued it. */
   if (state->db != NULL) {
+    Task *task =
+        Task_alloc(task_entry, TASK_STATUS_QUEUED, get_db_client_id(state->db));
     #if !RAY_USE_NEW_GCS
-      Task *task =
-          Task_alloc(task_entry, TASK_STATUS_QUEUED, get_db_client_id(state->db));
-
       if (from_global_scheduler) {
         /* If the task is from the global scheduler, it's already been added to
          * the task table, so just update the entry. */
@@ -924,12 +908,8 @@ std::list<TaskExecutionSpec>::iterator queue_task(
         task_table_add_task(state->db, task, NULL, NULL, NULL);
       }
     #else
-      TaskSpec* spec = task_entry.Spec();
-      auto data = MakeTaskTableData(task_entry, get_db_client_id(state->db), SchedulingState_QUEUED);
-      RAY_CHECK_OK(state->gcs_client.task_table().Add(ray::JobID::nil(), TaskSpec_task_id(spec), data,
-                                                      [](gcs::AsyncGcsClient *client,
-                                                         const TaskID &id,
-                                                         std::shared_ptr<TaskTableDataT> data) {}));
+      RAY_CHECK_OK(TaskTableAdd(&state->gcs_client, task));
+      Task_free(task);
     #endif
   }
 
@@ -1068,9 +1048,8 @@ void give_task_to_local_scheduler(LocalSchedulerState *state,
   CHECK(state->db != NULL);
   /* Assign the task to the relevant local scheduler. */
   DCHECK(state->config.global_scheduler_exists);
+  Task *task = Task_alloc(execution_spec, TASK_STATUS_SCHEDULED, local_scheduler_id);
   #if !RAY_USE_NEW_GCS
-    Task *task =
-        Task_alloc(execution_spec, TASK_STATUS_SCHEDULED, local_scheduler_id);
     auto retryInfo = RetryInfo{
         .num_retries = 0,  // This value is unused.
         .timeout = 0,      // This value is unused.
@@ -1079,12 +1058,8 @@ void give_task_to_local_scheduler(LocalSchedulerState *state,
 
     task_table_add_task(state->db, task, &retryInfo, NULL, state);
   #else
-    TaskSpec* spec = execution_spec.Spec();
-    auto data = MakeTaskTableData(execution_spec, local_scheduler_id, SchedulingState_SCHEDULED);
-    RAY_CHECK_OK(state->gcs_client.task_table().Add(ray::JobID::nil(), TaskSpec_task_id(spec), data,
-                                                    [](gcs::AsyncGcsClient *client,
-                                                       const TaskID &id,
-                                                       std::shared_ptr<TaskTableDataT> data) {}));
+    RAY_CHECK_OK(TaskTableAdd(&state->gcs_client, task));
+    Task_free(task);
   #endif
 }
 
@@ -1124,9 +1099,9 @@ void give_task_to_global_scheduler(LocalSchedulerState *state,
    * scheduler.
    */
   execution_spec.IncrementSpillbackCount();
-  #if !RAY_USE_NEW_GCS
-    Task *task =
+  Task *task =
         Task_alloc(execution_spec, TASK_STATUS_WAITING, DBClientID::nil());
+  #if !RAY_USE_NEW_GCS
     DCHECK(state->db != NULL);
     auto retryInfo = RetryInfo{
         .num_retries = 0,  // This value is unused.
@@ -1135,12 +1110,8 @@ void give_task_to_global_scheduler(LocalSchedulerState *state,
     };
     task_table_add_task(state->db, task, &retryInfo, NULL, state);
   #else
-    TaskSpec* spec = execution_spec.Spec();
-    auto data = MakeTaskTableData(execution_spec, DBClientID::nil(), SchedulingState_WAITING);
-    RAY_CHECK_OK(state->gcs_client.task_table().Add(ray::JobID::nil(), TaskSpec_task_id(spec), data,
-                                                    [](gcs::AsyncGcsClient *client,
-                                                       const TaskID &id,
-                                                       std::shared_ptr<TaskTableDataT> data) {}));
+    RAY_CHECK_OK(TaskTableAdd(&state->gcs_client, task));
+    Task_free(task);
   #endif
 }
