+++ src/local_scheduler/local_scheduler_shared.h
@@ -136,6 +136,4 @@ struct LocalSchedulerClient {
  */
 void LocalSchedulerState_free(LocalSchedulerState *state);
 
-std::shared_ptr<TaskTableDataT> MakeTaskTableData(const TaskExecutionSpec &execution_spec, const DBClientID& local_scheduler_id, SchedulingState scheduling_state);
-
 #endif /* LOCAL_SCHEDULER_SHARED_H */
